fix: support zod 4 only
