+++ .prettierignore
@@ -1,0 +1,1 @@
+__snapshots__
