+++ src/parse-env.test.ts
@@ -1,11 +1,11 @@
-import * as z from "zod/v3";
+import * as z from "zod";
 
 import { parseEnv } from "./index.js";
 import { port } from "./extra-schemas.js";
 
 // FIXME: many of these don't need to be part of parseCore tests, or at minimum
 // can be categorized further
-describe("parseCore v3", () => {
+describe("parseCore", () => {
   it("handles a basic case", () => {
     const x = parseEnv(
       {
@@ -480,7 +480,7 @@ describe("parseCore v3", () => {
     ).toThrow();
   });
 
-  const schemasWithDefaults: [z.ZodTypeAny, any][] = [
+  const schemasWithDefaults: [z.ZodType, any][] = [
     [z.number(), 5],
     [z.object({ a: z.string(), b: z.bigint() }), { a: "ok", b: 4n }],
   ];
