+++ src/parse-env.ts
@@ -1,26 +1,19 @@
-import * as z from "zod/v3";
-
+import { $ZodTypes, $ZodType, $ZodDefault } from "zod/v4/core";
 import { getSchemaWithPreprocessor } from "./preprocessors.js";
 import {
+  ErrorWithContext,
   makeDefaultReporter,
+  Reporter,
+  TokenFormatters,
   errorMap,
-  type TokenFormatters,
-  type ErrorWithContext,
-  type Reporter,
 } from "./reporter.js";
-
 import { resolveDefaultValueForSpec } from "./shared/utils.js";
+import type * as z from "zod/v4";
 import type { DeepReadonlyObject } from "./util/type-helpers.js";
 
-export type SimpleSchema<TOut = any, TIn = any> = z.ZodType<
-  TOut,
-  z.ZodTypeDef,
-  TIn
->;
+export type SimpleSchema<Out = any, In = any> = $ZodType<Out, In>;
 
-export type DetailedSpec<
-  TSchema extends SimpleSchema = SimpleSchema<unknown, unknown>,
-> =
+export type DetailedSpec<TSchema extends SimpleSchema = SimpleSchema> =
   TSchema extends SimpleSchema<any, infer TIn>
     ? {
         /**
@@ -32,6 +25,7 @@ export type DetailedSpec<
         /**
          * A description of this env var that's provided as help text if the
          * passed value fails validation, or is required but missing.
+         * @deprecated use e.g. `z.string().meta({ description: 'Your description' })` instead
          */
         description?: string;
 
@@ -80,14 +74,6 @@ export type ParsedSchema<T extends Schemas> = T extends any
     }
   : never;
 
-/**
- * Mostly an internal convenience function for testing. Returns the input
- * parameter unchanged, but with the same inference used in `parseEnv` applied.
- */
-export const inferSchemas = <T extends Schemas & RestrictSchemas<T>>(
-  schemas: T,
-): T & RestrictSchemas<T> => schemas;
-
 export type ParseEnv = <T extends Schemas & RestrictSchemas<T>>(
   env: Record<string, string | undefined>,
   schemas: T,
@@ -108,14 +94,13 @@ export function parseEnvImpl<T extends Schemas & RestrictSchemas<T>>(
   schemas: T,
   reporterOrTokenFormatters: Reporter | TokenFormatters,
 ): DeepReadonlyObject<ParsedSchema<T>> {
+  const parsed: Record<string, unknown> = {} as DeepReadonlyObject<
+    ParsedSchema<T>
+  >;
   const reporter =
     typeof reporterOrTokenFormatters === "function"
       ? reporterOrTokenFormatters
       : makeDefaultReporter(reporterOrTokenFormatters);
-
-  const parsed: Record<string, unknown> = {} as DeepReadonlyObject<
-    ParsedSchema<T>
-  >;
 
   const errors: ErrorWithContext[] = [];
 
@@ -125,10 +110,11 @@ export function parseEnvImpl<T extends Schemas & RestrictSchemas<T>>(
     let defaultUsed = false;
     let defaultValue: unknown;
     try {
-      if (schemaOrSpec instanceof z.ZodType) {
-        if (envValue == null && schemaOrSpec instanceof z.ZodDefault) {
+      if (schemaOrSpec instanceof $ZodType) {
+        if (envValue == null && schemaOrSpec instanceof $ZodDefault) {
           defaultUsed = true;
-          defaultValue = schemaOrSpec._def.defaultValue();
+          const spec = schemaOrSpec._zod;
+          defaultValue = spec.def.defaultValue;
           // we "unwrap" the default value ourselves and pass it to the schema.
           // in the very unlikely case that the value isn't stable AND
           // validation fails, this ensures the default value we report is the
@@ -137,12 +123,13 @@ export function parseEnvImpl<T extends Schemas & RestrictSchemas<T>>(
           //  we invoked the default getter and got 0.7, and then ran the parser
           //  against a missing env var and it generated another default of 0.4,
           //  we'd report a default value that _should_ have passed.)
-          parsed[key] = schemaOrSpec.parse(defaultValue, { errorMap });
+          parsed[key] = (spec.def.innerType as z.ZodType).parse(defaultValue, {
+            error: errorMap,
+          });
         } else {
-          parsed[key] = getSchemaWithPreprocessor(schemaOrSpec).parse(
-            envValue,
-            { errorMap },
-          );
+          parsed[key] = getSchemaWithPreprocessor(
+            schemaOrSpec as $ZodTypes,
+          ).parse(envValue, { error: errorMap });
         }
       } else if (envValue == null) {
         [defaultUsed, defaultValue] = resolveDefaultValueForSpec(
@@ -151,22 +138,20 @@ export function parseEnvImpl<T extends Schemas & RestrictSchemas<T>>(
         );
 
         if (defaultUsed) {
-          parsed[key] = schemaOrSpec.schema.parse(defaultValue, { errorMap });
+          parsed[key] = (schemaOrSpec.schema as z.ZodType).parse(defaultValue);
         } else {
           // if there's no default, pass our envValue through the
           // schema-with-preprocessor (it's an edge case, but our schema might
           // accept `null`, and the preprocessor will convert `undefined` to
           // `null` for us).
-          parsed[key] = getSchemaWithPreprocessor(schemaOrSpec.schema).parse(
-            envValue,
-            { errorMap },
-          );
+          parsed[key] = getSchemaWithPreprocessor(
+            schemaOrSpec.schema as $ZodTypes,
+          ).parse(envValue, { error: errorMap });
         }
       } else {
-        parsed[key] = getSchemaWithPreprocessor(schemaOrSpec.schema).parse(
-          envValue,
-          { errorMap },
-        );
+        parsed[key] = getSchemaWithPreprocessor(
+          schemaOrSpec.schema as $ZodTypes,
+        ).parse(envValue, { error: errorMap });
       }
     } catch (e) {
       errors.push({
