+++ src/preprocessors.ts
@@ -1,5 +1,4 @@
-import * as z from "zod/v3";
-
+import * as z from "zod/v4";
 import { assertNever } from "./util/type-helpers.js";
 import {
   bigInt,
@@ -13,60 +12,56 @@ import {
   throwIfCurrentlyUnsupported,
   throwIfWillNeverBeSupported,
 } from "./shared/processing.js";
-
-const { ZodFirstPartyTypeKind: TypeName } = z;
+import type * as zCore from "zod/v4/core";
 
 /**
  * Given a Zod schema, returns a function that tries to convert a string (or
  * undefined!) to a valid input type for the schema.
  */
 export function getPreprocessorByZodType(
-  schema: z.ZodFirstPartySchemaTypes,
+  schema: zCore.$ZodTypes,
 ): (arg: string | undefined) => unknown {
-  const def = schema._def;
-  const { typeName } = def;
+  const { def } = schema._zod;
 
-  switch (typeName) {
-    case TypeName.ZodString:
-    case TypeName.ZodEnum:
-    case TypeName.ZodUndefined:
+  switch (def.type) {
+    case "pipe":
+      return getPreprocessorByZodType(def.in as zCore.$ZodTypes);
+    case "string":
+    case "enum":
+    case "undefined":
       return identity;
 
-    case TypeName.ZodNumber:
+    case "number":
       return number;
 
-    case TypeName.ZodBigInt:
+    case "bigint":
       return bigInt;
 
-    case TypeName.ZodBoolean:
+    case "boolean":
       return boolean;
 
-    case TypeName.ZodArray:
-    case TypeName.ZodObject:
-    case TypeName.ZodTuple:
-    case TypeName.ZodRecord:
-    case TypeName.ZodIntersection:
+    case "array":
+    case "object":
+    case "tuple":
+    case "record":
+    case "intersection":
       return json;
 
-    case TypeName.ZodEffects:
-      return getPreprocessorByZodType(def.schema);
+    case "default":
+      return getPreprocessorByZodType(def.innerType as zCore.$ZodTypes);
 
-    case TypeName.ZodDefault:
-      // eslint-disable-next-line unicorn/consistent-destructuring -- false positive
-      return getPreprocessorByZodType(def.innerType);
-
-    case TypeName.ZodOptional: {
+    case "optional": {
       const { innerType } = def;
-      const pp = getPreprocessorByZodType(innerType);
+      const pp = getPreprocessorByZodType(innerType as zCore.$ZodTypes);
       return (arg) => {
         if (arg === undefined) return arg;
         return pp(arg);
       };
     }
 
-    case TypeName.ZodNullable: {
+    case "nullable": {
       const { innerType } = def;
-      const pp = getPreprocessorByZodType(innerType);
+      const pp = getPreprocessorByZodType(innerType as zCore.$ZodTypes);
       return (arg) => {
         // coerce undefined to null.
         if (arg == null) return null;
@@ -74,59 +69,60 @@ export function getPreprocessorByZodType(
       };
     }
 
-    case TypeName.ZodDate:
+    case "date":
       return date;
 
-    case TypeName.ZodLiteral:
-      switch (typeof def.value) {
+    case "literal":
+      switch (typeof def.values?.[0]) {
         case "number":
           return getPreprocessorByZodType({
-            _def: { typeName: TypeName.ZodNumber },
-          } as z.ZodFirstPartySchemaTypes);
+            _zod: { def: { type: "number" } },
+          } as zCore.$ZodTypes);
         case "string":
           return getPreprocessorByZodType({
-            _def: { typeName: TypeName.ZodString },
-          } as z.ZodFirstPartySchemaTypes);
+            _zod: { def: { type: "string" } },
+          } as zCore.$ZodTypes);
         case "boolean":
           return getPreprocessorByZodType({
-            _def: { typeName: TypeName.ZodBoolean },
-          } as z.ZodFirstPartySchemaTypes);
+            _zod: { def: { type: "boolean" } },
+          } as zCore.$ZodTypes);
         default:
           return (arg) => arg;
       }
 
-    case TypeName.ZodNull:
+    case "null":
       return nullProcessor;
 
-    case TypeName.ZodDiscriminatedUnion:
-    case TypeName.ZodUnion:
-    case TypeName.ZodNativeEnum:
-      return throwIfCurrentlyUnsupported(typeName);
+    case "union":
+      return throwIfCurrentlyUnsupported(def.type);
 
-    case TypeName.ZodAny:
-    case TypeName.ZodUnknown:
-      return throwIfUnknown(typeName);
+    case "any":
+    case "unknown":
+      return throwIfUnknown(def.type);
 
     // some of these types could maybe be supported (if only via the identity
     // function), but don't necessarily represent something meaningful as a
     // top-level schema passed to znv.
-    case TypeName.ZodVoid:
-    case TypeName.ZodNever:
-    case TypeName.ZodLazy:
-    case TypeName.ZodFunction:
-    case TypeName.ZodPromise:
-    case TypeName.ZodMap:
-    case TypeName.ZodSet:
-    case TypeName.ZodNaN:
-    case TypeName.ZodCatch:
-    case TypeName.ZodBranded:
-    case TypeName.ZodPipeline:
-    case TypeName.ZodSymbol:
-    case TypeName.ZodReadonly:
-      return throwIfWillNeverBeSupported(typeName);
-
+    case "success":
+    case "catch":
+    case "nan":
+    case "template_literal":
+    case "custom":
+    case "nonoptional":
+    case "prefault":
+    case "transform":
+    case "file":
+    case "void":
+    case "never":
+    case "lazy":
+    case "promise":
+    case "map":
+    case "set":
+    case "symbol":
+    case "readonly":
+      return throwIfWillNeverBeSupported(def.type);
     default: {
-      assertNever(typeName);
+      assertNever(def);
     }
   }
 }
@@ -135,7 +131,7 @@ export function getPreprocessorByZodType(
  * Given a Zod schema, return the schema wrapped in a preprocessor that tries to
  * convert a string to the schema's input type.
  */
-export function getSchemaWithPreprocessor(schema: z.ZodTypeAny) {
+export function getSchemaWithPreprocessor(schema: zCore.$ZodTypes) {
   return z.preprocess(
     getPreprocessorByZodType(schema) as (arg: unknown) => unknown,
     schema,
